Shutdown Dispatcher on spell check thread exit.
Normally, when a thread's ThreadStart finishes, a thread is terminated and cleaned up.  However, if a Dispatcher was created for that thread, the thread and Dispatcher-related resources will not be cleaned up.  One such resource is a window handle, and if enough of these are leaked, the process will run out of valid handles and crash.

In this case, the spell checker background thread is indirectly creating a Dispatcher, since it creates a WPF object (which creates a Dispatcher), and so needs to clean up the Dispatcher before the thread exits.

This was identified by the Visual Studio watson process as causing ~1500 crashes.