Multi-Language Support, Beta, still lots of bugs in the configuration window.
Implemented Features:

- Support for spell checking multiple languages.
   The checker checks for correct spelling in multiple configurable languages in the same text.
- Configurable language support through smarttags.
- Configuration of custom Dictionaries.

Spell checker now determines language on a per sentence base. It tolerates foreign language in a sentence, if it is a group of at least SpellingTagger.MinForeignWordSequence words which is currently set to 3 words.
If only one language is enabled performance is identical to the old algorithm, with more languages, it decreqases proportional to the misspells.